# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Trivial Scanner Dashboard API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/hosts": {
      "get": {
        "tags": [
          "Hostname"
        ],
        "summary": "Retrieve Hosts",
        "description": "Retrieves a collection of your own Trivial Scanner reports, providing\na distinct list of hosts and ports, optionally returning the latest host\nfull record",
        "operationId": "retrieve_hosts_hosts_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            },
            "name": "authorization",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "title": "X-Trivialscan-Account",
              "type": "string"
            },
            "name": "x-trivialscan-account",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Retrieve Hosts Hosts Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Host"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No scan data is present for this account"
          },
          "401": {
            "description": "Authorization Header was sent but something was not valid (check the logs), likely signed the wrong HTTP method or forgot to sign the base64 encoded POST data"
          },
          "403": {
            "description": "Authorization Header was not sent, or dropped at a proxy (requesters issue) or the CDN (that one is our server misconfiguration)"
          },
          "500": {
            "description": "An unhandled error occurred during an AWS request for data access"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/host/{hostname}": {
      "get": {
        "tags": [
          "Hostname"
        ],
        "summary": "Retrieve Host",
        "description": "Retrieves TLS data on any hostname, providing an optional port number",
        "operationId": "retrieve_host_host__hostname__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Hostname",
              "type": "string"
            },
            "name": "hostname",
            "in": "path"
          },
          {
            "description": "defaults to 443 when not searching by date, otherwise leaving the port empty returns first found on date regardless of port",
            "required": false,
            "schema": {
              "title": "Port",
              "type": "integer",
              "description": "defaults to 443 when not searching by date, otherwise leaving the port empty returns first found on date regardless of port"
            },
            "name": "port",
            "in": "query"
          },
          {
            "description": "Return the result for specific date rather than the latest (default) Host information. Represented in ISO 8601 format; 2008-09-15T15:53:00+05:00",
            "required": false,
            "schema": {
              "title": "Last Updated",
              "type": "string",
              "description": "Return the result for specific date rather than the latest (default) Host information. Represented in ISO 8601 format; 2008-09-15T15:53:00+05:00",
              "format": "date-time"
            },
            "name": "last_updated",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            },
            "name": "authorization",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "title": "X-Trivialscan-Account",
              "type": "string"
            },
            "name": "x-trivialscan-account",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostResponse"
                }
              }
            }
          },
          "204": {
            "description": "No scan data is present for this account"
          },
          "401": {
            "description": "Authorization Header was sent but something was not valid (check the logs), likely signed the wrong HTTP method or forgot to sign the base64 encoded POST data"
          },
          "403": {
            "description": "Authorization Header was not sent, or dropped at a proxy (requesters issue) or the CDN (that one is our server misconfiguration)"
          },
          "500": {
            "description": "An unhandled error occurred during an AWS request for data access"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/search/host/{hostname}": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Search Hostname",
        "description": "Search matching hostname, returning exact matches and knowm (scanned, if any) subdomains",
        "operationId": "search_hostname_search_host__hostname__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Hostname",
              "type": "string"
            },
            "name": "hostname",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            },
            "name": "authorization",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "title": "X-Trivialscan-Account",
              "type": "string"
            },
            "name": "x-trivialscan-account",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Search Hostname Search Host  Hostname  Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResult"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No search results matching this query"
          },
          "401": {
            "description": "Authorization Header was sent but something was not valid (check the logs), likely signed the wrong HTTP method or forgot to sign the base64 encoded POST data"
          },
          "403": {
            "description": "Authorization Header was not sent, or dropped at a proxy (requesters issue) or the CDN (that one is our server misconfiguration)"
          },
          "500": {
            "description": "An unhandled error occurred during an AWS request for data access"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Certificate": {
        "title": "Certificate",
        "required": [
          "sha1_fingerprint"
        ],
        "type": "object",
        "properties": {
          "authority_key_identifier": {
            "title": "Authority Key Identifier",
            "type": "string"
          },
          "expired": {
            "title": "Expired",
            "type": "boolean"
          },
          "expiry_status": {
            "title": "Expiry Status",
            "type": "string"
          },
          "extensions": {
            "title": "Extensions",
            "type": "array",
            "items": {},
            "default": []
          },
          "external_refs": {
            "title": "External Refs",
            "type": "object",
            "additionalProperties": {
              "maxLength": 65536,
              "minLength": 1,
              "type": "string",
              "format": "uri"
            },
            "default": {}
          },
          "is_self_signed": {
            "title": "Is Self Signed",
            "type": "boolean"
          },
          "issuer": {
            "title": "Issuer",
            "type": "string"
          },
          "known_compromised": {
            "title": "Known Compromised",
            "type": "boolean"
          },
          "md5_fingerprint": {
            "title": "Md5 Fingerprint",
            "type": "string"
          },
          "not_after": {
            "title": "Not After",
            "type": "string",
            "format": "date-time"
          },
          "not_before": {
            "title": "Not Before",
            "type": "string",
            "format": "date-time"
          },
          "public_key_curve": {
            "title": "Public Key Curve",
            "type": "string"
          },
          "public_key_exponent": {
            "title": "Public Key Exponent",
            "exclusiveMinimum": 0.0,
            "type": "integer"
          },
          "public_key_modulus": {
            "title": "Public Key Modulus",
            "exclusiveMinimum": 0.0,
            "type": "integer"
          },
          "public_key_size": {
            "title": "Public Key Size",
            "exclusiveMinimum": 0.0,
            "type": "integer"
          },
          "public_key_type": {
            "$ref": "#/components/schemas/PublicKeyType"
          },
          "revocation_crl_urls": {
            "title": "Revocation Crl Urls",
            "type": "array",
            "items": {
              "maxLength": 65536,
              "minLength": 1,
              "type": "string",
              "format": "uri"
            },
            "default": []
          },
          "san": {
            "title": "San",
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "serial_number": {
            "title": "Serial Number",
            "type": "string"
          },
          "serial_number_decimal": {
            "title": "Serial Number Decimal"
          },
          "serial_number_hex": {
            "title": "Serial Number Hex",
            "type": "string"
          },
          "sha1_fingerprint": {
            "title": "Sha1 Fingerprint",
            "type": "string"
          },
          "sha256_fingerprint": {
            "title": "Sha256 Fingerprint",
            "type": "string"
          },
          "signature_algorithm": {
            "title": "Signature Algorithm",
            "type": "string"
          },
          "spki_fingerprint": {
            "title": "Spki Fingerprint",
            "type": "string"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "subject_key_identifier": {
            "title": "Subject Key Identifier",
            "type": "string"
          },
          "validation_level": {
            "$ref": "#/components/schemas/ValidationLevel"
          },
          "validation_oid": {
            "title": "Validation Oid",
            "type": "string"
          },
          "version": {
            "title": "Version"
          },
          "type": {
            "$ref": "#/components/schemas/CertificateType"
          }
        }
      },
...
      "ClientInfo": {
        "title": "ClientInfo",
        "type": "object",
        "properties": {
          "operating_system": {
            "title": "Operating System",
            "type": "string"
          },
          "operating_system_release": {
            "title": "Operating System Release",
            "type": "string"
          },
          "operating_system_version": {
            "title": "Operating System Version",
            "type": "string"
          },
          "architecture": {
            "title": "Architecture",
            "type": "string"
          }
        }
      },
...
      "Flags": {
        "title": "Flags",
        "type": "object",
        "properties": {
          "hide_progress_bars": {
            "title": "Hide Progress Bars",
            "type": "boolean"
          },
          "synchronous_only": {
            "title": "Synchronous Only",
            "type": "boolean"
          },
          "hide_banner": {
            "title": "Hide Banner",
            "type": "boolean"
          },
          "track_changes": {
            "title": "Track Changes",
            "type": "boolean"
          },
          "previous_report": {
            "title": "Previous Report",
            "type": "string"
          },
          "quiet": {
            "title": "Quiet",
            "type": "boolean"
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "Host": {
        "title": "Host",
        "required": [
          "transport"
        ],
        "type": "object",
        "properties": {
          "last_updated": {
            "title": "Last Updated",
            "type": "string",
            "format": "date-time"
          },
          "transport": {
            "$ref": "#/components/schemas/HostTransport"
          },
          "tls": {
            "$ref": "#/components/schemas/HostTLS"
          },
          "http": {
            "title": "Http",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HostHTTP"
            }
          },
          "monitoring_enabled": {
            "title": "Monitoring Enabled",
            "type": "boolean",
            "default": false
          }
        }
      },
      "HostHTTP": {
        "title": "HostHTTP",
        "required": [
          "title",
          "status_code",
          "headers",
          "body_hash"
        ],
        "type": "object",
        "properties": {
          "title": {
            "title": "Title",
            "type": "string"
          },
          "status_code": {
            "title": "Status Code",
            "maximum": 599.0,
            "minimum": 100.0,
            "type": "integer"
          },
          "headers": {
            "title": "Headers",
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "body_hash": {
            "title": "Body Hash",
            "type": "string"
          }
        }
      },
      "HostResponse": {
        "title": "HostResponse",
        "required": [
          "host",
          "reports",
          "versions",
          "external_refs"
        ],
        "type": "object",
        "properties": {
          "host": {
            "$ref": "#/components/schemas/Host"
          },
          "reports": {
            "title": "Reports",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReportSummary"
            }
          },
          "versions": {
            "title": "Versions",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "external_refs": {
            "title": "External Refs",
            "type": "object",
            "additionalProperties": {
              "anyOf": [
                {
                  "maxLength": 65536,
                  "minLength": 1,
                  "type": "string",
                  "format": "uri"
                },
                {
                  "type": "string"
                }
              ]
            }
          }
        }
      },
      "HostTLS": {
        "title": "HostTLS",
        "required": [
          "client",
          "cipher",
          "protocol",
          "session_resumption"
        ],
        "type": "object",
        "properties": {
          "certificates": {
            "title": "Certificates",
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "client": {
            "$ref": "#/components/schemas/HostTLSClient"
          },
          "cipher": {
            "$ref": "#/components/schemas/HostTLSCipher"
          },
          "protocol": {
            "$ref": "#/components/schemas/HostTLSProtocol"
          },
          "session_resumption": {
            "$ref": "#/components/schemas/HostTLSSessionResumption"
          }
        }
      },
      "HostTLSCipher": {
        "title": "HostTLSCipher",
        "required": [
          "offered",
          "offered_rfc",
          "negotiated",
          "negotiated_bits",
          "negotiated_rfc"
        ],
        "type": "object",
        "properties": {
          "forward_anonymity": {
            "title": "Forward Anonymity",
            "type": "boolean",
            "default": false
          },
          "offered": {
            "title": "Offered",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "offered_rfc": {
            "title": "Offered Rfc",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "negotiated": {
            "title": "Negotiated",
            "type": "string"
          },
          "negotiated_bits": {
            "title": "Negotiated Bits",
            "exclusiveMinimum": 0.0,
            "type": "integer"
          },
          "negotiated_rfc": {
            "title": "Negotiated Rfc",
            "type": "string"
          }
        }
      },
      "HostTLSClient": {
        "title": "HostTLSClient",
        "type": "object",
        "properties": {
          "certificate_mtls_expected": {
            "title": "Certificate Mtls Expected",
            "type": "boolean",
            "default": false
          },
          "certificate_trusted": {
            "title": "Certificate Trusted",
            "type": "boolean",
            "default": false
          },
          "certificate_match": {
            "title": "Certificate Match",
            "type": "boolean",
            "default": false
          },
          "expected_client_subjects": {
            "title": "Expected Client Subjects",
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          }
        }
      },
      "HostTLSProtocol": {
        "title": "HostTLSProtocol",
        "required": [
          "negotiated",
          "preferred",
          "offered"
        ],
        "type": "object",
        "properties": {
          "negotiated": {
            "title": "Negotiated",
            "type": "string"
          },
          "preferred": {
            "title": "Preferred",
            "type": "string"
          },
          "offered": {
            "title": "Offered",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "HostTLSSessionResumption": {
        "title": "HostTLSSessionResumption",
        "required": [
          "cache_mode",
          "tickets",
          "ticket_hint"
        ],
        "type": "object",
        "properties": {
          "cache_mode": {
            "title": "Cache Mode",
            "type": "string"
          },
          "tickets": {
            "title": "Tickets",
            "type": "boolean"
          },
          "ticket_hint": {
            "title": "Ticket Hint",
            "type": "boolean"
          }
        }
      },
      "HostTransport": {
        "title": "HostTransport",
        "required": [
          "hostname"
        ],
        "type": "object",
        "properties": {
          "error": {
            "title": "Error",
            "maxItems": 2,
            "minItems": 2,
            "type": "array",
            "items": [
              {
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          },
          "hostname": {
            "title": "Domain Name",
            "type": "string"
          },
          "port": {
            "title": "Port",
            "exclusiveMinimum": 0.0,
            "type": "integer",
            "default": 443
          },
          "sni_support": {
            "title": "Sni Support",
            "type": "boolean"
          },
          "peer_address": {
            "title": "Peer Address",
            "type": "string",
            "format": "ipvanyaddress"
          },
          "certificate_mtls_expected": {
            "title": "Certificate Mtls Expected",
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "ReportSummary": {
        "title": "ReportSummary",
        "required": [
          "report_id"
        ],
        "type": "object",
        "properties": {
          "generator": {
            "title": "Generator",
            "type": "string",
            "default": "trivialscan"
          },
          "version": {
            "title": "Version",
            "type": "string",
            "description": "trivialscan CLI version"
          },
          "account_name": {
            "title": "Account Name",
            "type": "string",
            "description": "Trivial Security account name"
          },
          "client_name": {
            "title": "Client Name",
            "type": "string",
            "description": "Machine name where trivialscan CLI execcutes"
          },
          "report_id": {
            "title": "Report Id",
            "type": "string"
          },
          "project_name": {
            "title": "Project Name",
            "type": "string"
          },
          "targets": {
            "title": "Targets",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Host"
            },
            "default": []
          },
          "date": {
            "title": "Date",
            "type": "string",
            "format": "date-time"
          },
          "execution_duration_seconds": {
            "title": "Execution Duration Seconds",
            "exclusiveMinimum": 0.0,
            "type": "number"
          },
          "score": {
            "title": "Score",
            "type": "integer",
            "default": 0
          },
          "results": {
            "title": "Results",
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "certificates": {
            "title": "Certificates",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Certificate"
            },
            "default": []
          },
          "results_uri": {
            "title": "Results Uri",
            "type": "string"
          },
          "flags": {
            "$ref": "#/components/schemas/Flags"
          },
          "config": {
            "$ref": "#/components/schemas/Config"
          },
          "client": {
            "$ref": "#/components/schemas/ClientInfo"
          },
          "type": {
            "$ref": "#/components/schemas/ScanRecordType"
          },
          "category": {
            "$ref": "#/components/schemas/ScanRecordCategory"
          },
          "is_passive": {
            "title": "Is Passive",
            "type": "boolean",
            "default": true
          }
        }
      },
...
      "ScanRecordType": {
        "title": "ScanRecordType",
        "enum": [
          "Managed Monitoring",
          "Managed On-demand",
          "Customer-managed"
        ],
        "type": "string",
        "description": "An enumeration."
      },
...
      "SearchResult": {
        "title": "SearchResult",
        "required": [
          "ip_addr"
        ],
        "type": "object",
        "properties": {
          "last_scanned": {
            "title": "Last Scanned",
            "type": "integer"
          },
          "hostname": {
            "title": "Hostname",
            "type": "string"
          },
          "monitoring": {
            "title": "Monitoring",
            "type": "boolean",
            "default": false
          },
          "queued_timestamp": {
            "title": "Queued Timestamp",
            "type": "integer"
          },
          "queue_status": {
            "title": "Queue Status",
            "type": "string",
            "default": false
          },
          "ip_addr": {
            "title": "Ip Addr",
            "type": "array",
            "items": {
              "type": "string",
              "format": "ipvanyaddress"
            }
          },
          "resolved_ip": {
            "title": "Resolved Ip",
            "type": "array",
            "items": {
              "type": "string",
              "format": "ipvanyaddress"
            }
          },
          "ports": {
            "title": "Ports",
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "reports": {
            "title": "Reports",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "scanned": {
            "title": "Scanned",
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      },
...
    }
  }
}